Simply bleak scanner adapter + refactor some code
The event are now emitted when a device is inserted / update in the
repository. For that we use an observer that will watch the state of the
repo. Those interested observers (here the event bus that will publish
the update on the websocket) will be notified when a change occurred.
This simplify the logic in the scanner but also defines clear
boundaries, as each module as well-defined responsabilities.

Signed-off-by: Lionel H <me@nullbyte.be>